Switch to `ioredis` and `redlock`
The `redlock` package provides better guarantees comparing to `redis-lock` plus better control over the lock which allows to handle the lock expiration case properly.
The `ioredis` package has been brought mainly because `redlock` requires it but it also allows for simpler use of the library as connecting the client before usage is not necessary anymore.